services/feedback-pipeline/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sentence-transformers>=3.0.0
supabase>=2.4.0
groq>=0.5.0
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.6.0
numpy>=1.26.0
scikit-learn>=1.4.0
joblib>=1.3.0
shap>=0.45.0
bertopic>=0.16.0
umap-learn>=0.5.5
hdbscan>=0.8.33
matplotlib>=3.8.0
mlflow>=2.12.0
pandas>=2.2.0
scipy>=1.12.0
transformers>=4.38.0
